requirements_fixed.txt
# Core AutoGen and AI dependencies
autogen-agentchat>=0.6.0
groq>=0.25.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
docx2txt>=0.8
openpyxl>=3.1.0
pandas>=2.0.0

# Vector database for knowledge base
chromadb>=1.0.0

# Web scraping and URL processing
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
urllib3>=2.0.0
certifi>=2023.7.22

# Web UI
streamlit>=1.40.0
streamlit-chat>=0.1.0

# HTTP client for async requests
httpx>=0.25.0

# MCP Server dependencies (if needed)
# mcp>=0.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Additional utility packages
rich>=13.0.0  # For better terminal output
typing-extensions>=4.0.0  # For enhanced type hints
